Configures Amplify for authentication
Introduces Amplify configuration using environment variables for Cognito integration.
This centralizes the configuration and allows for easier management of authentication settings.

Additionally, this commit updates the authentication flow to properly handle user sessions and logout functionality. It ensures that user tokens are cleared from cookies upon logout and that the auth store is reset, providing a cleaner user experience. Also add a null check for user email when logging access.