Adds a Cognito Local dashboard
Adds a SvelteKit dashboard to list and administer users
in a local Cognito pool, due to the absence of a built-in
graphical interface in cognito-local.

This includes creating, listing, and deleting users through
API calls to the local Cognito instance.

Also saves endpoint and user pool ID in local storage.